Updated commands architecture
